requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database connections
neo4j==5.14.0
psycopg2-binary==2.9.9

# AI and ML
ollama==0.1.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.0

# Optional: For enhanced text processing
nltk==3.8.1
scikit-learn==1.3.2